@@ -26,7 +26,7 @@
         }
     },
     "require": {
-        "php": "^8.0",
+        "php": "^8.1",
         "ext-json": "*",
         "doctrine/inflector": "^1.4 || ^2.0",
         "jane-php/json-schema-runtime": "^7.5",
