@@ -7,30 +7,18 @@ use PhpParser\Node;
 /**
  * File generated.
  *
- * Simply a correspondance between a filename and a AST
+ * Simply a correspondence between a filename and a AST
  */
 class File
 {
-    /**
-     * Relative path of the file generated.
-     */
-    private $filename;
-
-    /**
-     * Ast generated.
-     */
-    private $node;
-
-    /**
-     * Type of generation (model / normalizer / ...).
-     */
-    private $type;
-
-    public function __construct(string $filename, Node $node, string $type)
-    {
-        $this->filename = $filename;
-        $this->node = $node;
-        $this->type = $type;
+    public function __construct(
+        /** Relative path of the file generated */
+        private readonly string $filename,
+        /** Ast generated */
+        private readonly Node $node,
+        /** Type of generation (model / normalizer / ...) */
+        private readonly string $type,
+    ) {
     }
 
     public function getFilename(): string
