@@ -7,10 +7,8 @@ use Jane\Component\JsonSchema\Registry\Registry;
 
 class ChainGuesser implements TypeGuesserInterface, PropertiesGuesserInterface, ClassGuesserInterface
 {
-    /**
-     * @var GuesserInterface[]
-     */
-    private $guessers = [];
+    /** @var array<GuesserInterface> */
+    private array $guessers = [];
 
     public function addGuesser(GuesserInterface $guesser): void
     {
@@ -21,9 +19,6 @@ class ChainGuesser implements TypeGuesserInterface, PropertiesGuesserInterface, 
         $this->guessers[] = $guesser;
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function guessClass($object, string $name, string $reference, Registry $registry): void
     {
         foreach ($this->guessers as $guesser) {
@@ -37,9 +32,6 @@ class ChainGuesser implements TypeGuesserInterface, PropertiesGuesserInterface, 
         }
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function guessType($object, string $name, string $reference, Registry $registry): Type
     {
         $type = null;
@@ -61,9 +53,6 @@ class ChainGuesser implements TypeGuesserInterface, PropertiesGuesserInterface, 
         return $type;
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function guessProperties($object, string $name, string $reference, Registry $registry): array
     {
         $properties = [];
