@@ -13,24 +13,18 @@ use PhpParser\Node\Scalar;
 
 class ObjectType extends Type
 {
-    private $className;
-
-    private $namespace;
-
-    private $discriminants;
-
-    public function __construct(object $object, string $className, string $namespace, array $discriminants = [])
-    {
+    /**
+     * @param array<string, array<string>|null> $discriminants
+     */
+    public function __construct(
+        object $object,
+        private readonly string $className,
+        private readonly string $namespace,
+        private readonly array $discriminants = [],
+    ) {
         parent::__construct($object, 'object');
-
-        $this->namespace = $namespace;
-        $this->className = $className;
-        $this->discriminants = $discriminants;
     }
 
-    /**
-     * ({@inheritDoc}.
-     */
     protected function createDenormalizationValueStatement(Context $context, Expr $input, bool $normalizerFromObject = true): Expr
     {
         $denormalizerVar = new Expr\PropertyFetch(new Expr\Variable('this'), 'denormalizer');
@@ -49,9 +43,6 @@ class ObjectType extends Type
         ]);
     }
 
-    /**
-     * ({@inheritDoc}.
-     */
     protected function createNormalizationValueStatement(Context $context, Expr $input, bool $normalizerFromObject = true): Expr
     {
         $normalizerVar = new Expr\PropertyFetch(new Expr\Variable('this'), 'normalizer');
@@ -66,9 +57,6 @@ class ObjectType extends Type
         ]);
     }
 
-    /**
-     * ({@inheritDoc}.
-     */
     public function createConditionStatement(Expr $input): Expr
     {
         $conditionStatement = parent::createConditionStatement($input);
@@ -112,21 +100,15 @@ class ObjectType extends Type
         return $conditionStatement;
     }
 
-    /**
-     * ({@inheritDoc}.
-     */
-    public function getTypeHint(string $currentNamespace): Name
+    public function getTypeHint(string $namespace): Name
     {
-        if ('\\' . $currentNamespace . '\\' . $this->className === $this->getFqdn()) {
+        if ('\\' . $namespace . '\\' . $this->className === $this->getFqdn()) {
             return new Name($this->className);
         }
 
         return new Name($this->getFqdn());
     }
 
-    /**
-     * ({@inheritDoc}.
-     */
     public function getDocTypeHint(string $namespace): Name
     {
         return $this->getTypeHint($namespace);
